Revert "fix(qualification): fail closed on incomplete evidence"
This reverts commit df6d1bad4b8c87602ba17bda774729c9846f2421.
