+++ docs/index.mdx
@@ -78,6 +78,45 @@ By default, NemoClaw installs the OpenClaw agent. Use one of the following quick
 | LangChain Deep Agents Code | [Quickstart with Deep Agents](../deepagents/get-started/quickstart) |
 | Pi (release candidate) | [Quickstart with Pi](../pi/get-started/quickstart) |
 
+<section>
+
+## NemoClaw Docs for AI Agents
+
+Use NemoClaw's Markdown docs or the docs-routing skill when you want your AI coding agent to help you install or operate a sandbox.
+The assistant fetches the same pages that power this site and applies them to your local environment.
+
+<Cards>
+
+<Card title="Add NemoClaw Docs to Your Agent" href="resources/agent-skills">
+
+Give your assistant the docs entry points or optional routing skill so NemoClaw guidance is available directly in chat.
+
+<Badge intent="tip" minimal outlined>Resource</Badge>
+</Card>
+
+</Cards>
+
+</section>
+
+<section>
+
+## Resources
+
+Use the following resources to explore NemoClaw end-to-end examples, showcases, and integrations.
+
+<Cards>
+
+<Card title="Explore Community Solutions" href="resources/community-contributions">
+
+Learn where to contribute third-party solutions, showcases, and integrations built with NemoClaw.
+
+<Badge intent="tip" minimal outlined>Examples</Badge>
+</Card>
+
+</Cards>
+
+</section>
+
 ## Select User Guide Variant for Your Preferred Agent
 
 Select the user guide variant for your preferred agent to sandbox. You can also switch guide variants from the dropdown at the top of the left navigation pane.
